Rename auth to googleAuth
