+++ components/layout/staff-layout.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React, { ReactNode } from 'react';
 import { useUser } from '../../lib/contexts/user-context';
-import { signOut } from '../../lib/utils/auth';
+import { signOut } from '../../lib/utils/googleAuth';
 import { hasPhaseBanner } from '../../lib/utils/phase-banner';
 import Header from '../header';
 import PhaseBanner from '../phase-banner';
