+++ pages/applications/view/[id]/[person]/index.tsx
@@ -9,7 +9,7 @@ import {
   getRedirect,
   getSession,
   HackneyGoogleUserWithPermissions,
-} from '../../../../../lib/utils/auth';
+} from '../../../../../lib/utils/googleAuth';
 import Custom404 from '../../../../404';
 import CheckBoxList, {
   CheckBoxListPageProps,
